Refactor: Split accounts into different files and adjust naming
